feat: move to generic captcha plugin, supporting multiple providers
