--- docs/content/docs/plugins/cloudflare-turnstile.mdx
@@ -1,81 +1,0 @@
----
-title: Cloudflare Turnstile
-description: Cloudflare Turnstile plugin
----
-
-The **Cloudflare Turnstile Plugin** integrates bot protection into your Better Auth system by adding CAPTCHA verification for key endpoints. This plugin ensures that only human users can perform actions like signing up, signing in, or resetting passwords.
-
-## Installation
-
-<Steps>
-  <Step>
-    ### Add the plugin to your **auth** config
-
-    ```ts title="auth.ts"
-    import { betterAuth } from "better-auth";
-    import { turnstile } from "better-auth/plugins";
-
-    export const auth = betterAuth({
-        plugins: [ // [!code highlight]
-            turnstile({ // [!code highlight]
-                secretKey: process.env.TURNSTILE_SECRET_KEY!, // [!code highlight]
-            }), // [!code highlight]
-        ], // [!code highlight]
-    });
-    ```
-
-  </Step>
-  <Step>
-    ### Add the CAPTCHA token to your request headers
-
-    Add the CAPTCHA token to your request headers for all protected endpoints. This example shows how to include it in a `signIn` request:
-
-    ```ts
-    await authClient.signIn.email({
-        email: "user@example.com",
-        password: "secure-password",
-        fetchOptions: { // [!code highlight]
-            headers: { // [!code highlight]
-                "x-turnstile-captcha-response": turnstileToken, // [!code highlight]
-            }, // [!code highlight]
-        },
-    });
-    ```
-
-    To implement Cloudflare Turnstile on the client side, see the official [Cloudflare Turnstile documentation](https://developers.cloudflare.com/turnstile/) or use a library like [react-turnstile](https://www.npmjs.com/package/@marsidev/react-turnstile).
-
-  </Step>
-</Steps>
-
-## How it works
-
-<Steps>
-  <Step>
-    The plugin acts as a middleware: it intercepts all `POST` requests to configured endpoints (see `endpoints`
-    in the [Plugin Options](#plugin-options) section).
-  </Step>
-  <Step>
-    it validates the CAPTCHA token on the server, by calling Turnstile's `/siteverify`.
-  </Step>
-  <Step>
-    - if the token is missing, gets rejected by Turnstile, or if the `/siteverify` endpoint is
-    unavailable, the plugin returns an error and interrupts the request.
-    - if the token is accepted by Turnstile, the request is allowed to proceed.
-
-  </Step>
-</Steps>
-
-## Plugin Options
-
-- **`secretKey` (required)**
-  Your Cloudflare Turnstile secret key used for the server-side validation of CAPTCHA tokens.
-- **`endpoints` (optional)**
-  An array of paths where CAPTCHA validation is enforced. Defaults to:
-  ```ts
-  ["/sign-up", "/sign-in", "/forget-password"];
-  ```
-- **`siteVerifyURL` (optional)**
-  The URL for the CAPTCHA verification request. Defaults to:
-  ```
-  https://challenges.cloudflare.com/turnstile/v0/siteverify
-  ```
