+++ packages/better-auth/src/plugins/captcha/error-codes.ts
@@ -1,0 +1,6 @@
+export const CAPTCHA_ERROR_CODES = {
+  MISSING_RESPONSE: "Missing CAPTCHA response",
+  SERVICE_UNAVAILABLE: "CAPTCHA service unavailable",
+  VERIFICATION_FAILED: "Captcha verification failed",
+  UNKNOWN_ERROR: "Something went wrong",
+} as const;
