+++ packages/better-auth/src/plugins/captcha/verify-handlers/index.ts
@@ -1,0 +1,2 @@
+export { cloudflareTurnstile } from "./cloudflare-turnstile";
+export { googleReCAPTCHA } from "./google-recaptcha";
