--- packages/better-auth/src/plugins/cloudflare-turnstile/constants.ts
@@ -1,4 +1,0 @@
-export const defaultEndpoints = ["/sign-up", "/sign-in", "/forget-password"];
-
-export const defaultSiteVerify =
-	"https://challenges.cloudflare.com/turnstile/v0/siteverify";
