--- packages/better-auth/src/plugins/cloudflare-turnstile/error-codes.ts
@@ -1,6 +1,0 @@
-export const TURNSTILE_ERROR_CODES = {
-	MISSING_CAPTCHA_RESPONSE: "Missing CAPTCHA response",
-	CAPTCHA_SERVICE_UNAVAILABLE: "CAPTCHA service unavailable",
-	CAPTCHA_VERIFICATION_FAILED: "Captcha verification failed",
-	UNKNOWN_ERROR: "Something went wrong",
-} as const;
