--- packages/better-auth/src/plugins/cloudflare-turnstile/types.ts
@@ -1,18 +1,0 @@
-export type TurnstileSiteVerifyResponse =
-	| {
-			success: true;
-			"error-codes"?: string[];
-			challenge_ts?: string;
-			hostname?: string;
-			action?: string;
-			cdata?: string;
-			metadata?: {
-				interactive: boolean;
-			};
-			messages?: string[];
-	  }
-	| {
-			success: false;
-			"error-codes": string[];
-			messages: string[];
-	  };
